requirements.txt
requests
pandas
psycopg2
python-dotenv